Add HRESULT type and buffer changes.
- Rename NativeBuffer to HeapBuffer
- Add ISizedBuffer and IBuffer interfaces
- Rename NativeBufferReader to Reader and move size checks
   to  derived CheckedReader.
- Add abstract SizedBuffer class
- Add IBuffer to SafeHandle

+++ src/WInterop/ErrorHandling/DataTypes/DriveNotReadyException.cs
@@ -13,13 +13,13 @@ namespace WInterop.ErrorHandling.DataTypes
     public class DriveNotReadyException : IOException
     {
         public DriveNotReadyException()
-            : base() { HResult = ErrorMacros.HRESULT_FROM_WIN32(WindowsError.ERROR_NOT_READY); }
+            : base() { HResult = (int)ErrorMacros.HRESULT_FROM_WIN32(WindowsError.ERROR_NOT_READY); }
 
         public DriveNotReadyException(string message)
-            : base(message) { HResult = ErrorMacros.HRESULT_FROM_WIN32(WindowsError.ERROR_NOT_READY); }
+            : base(message) { HResult = (int)ErrorMacros.HRESULT_FROM_WIN32(WindowsError.ERROR_NOT_READY); }
 
         public DriveNotReadyException(string message, Exception innerException)
-            : base(message, innerException) { HResult = ErrorMacros.HRESULT_FROM_WIN32(WindowsError.ERROR_NOT_READY); }
+            : base(message, innerException) { HResult = (int)ErrorMacros.HRESULT_FROM_WIN32(WindowsError.ERROR_NOT_READY); }
 
         public DriveNotReadyException(string message, int hresult)
             : base(message, hresult) { }
